requirements.txt
requests>=2.31.0
python-dotenv>=1.0.0
typing-extensions>=4.9.0